+++ packages/expo-cli/src/commands/webhooks/webhooksRemoveAsync.ts
@@ -1,0 +1,14 @@
+import assert from 'assert';
+
+import { setupAsync } from './utils';
+
+type Options = {
+  id?: string;
+};
+
+export async function actionAsync(projectRoot: string, { id }: Options) {
+  assert(typeof id === 'string', '--id must be a webhook ID');
+  const { project, client } = await setupAsync(projectRoot);
+
+  await client.deleteAsync(`projects/${project.id}/webhooks/${id}`);
+}
